Add files via upload
Updating all of the code.  Added on new thread creation methods - these methods now only create 5 threads that never terminate, and continue running and updating what number to check.  This was required because creating new threads when one died did not work and ended up having one thread running or 17.  Comments are yet to be added.